Write results to STDOUT if warn-only is passed
The Code Climate platforms requires a successful analysis run to exit
with code 0 and write its results to STDOUT and a failed engine run to
exit with code 1 and write debug information to STDERR. A recent
refactoring to exit code and output handling changed this behavior such
that errors could be written to STDOUT and a successful analysis could
exit with code 1 if --warn-only wasn't passed.

This changes nsp behavior to always write vulnerability results to
STDOUT and exit with code 0 if --warn-only is passed. It also ensures
that if an error is raised, output will be written to STDERR and nsp
exits with code 1.

* Edits to brace formatting to respond to ESLint feedback