Return error message in Code Climate runs
Stringifying an Error object results in {} so the formatter was not
returning the actual underlying error with the Code Climate formatter.
This change returns the error so that they user can take action if an
analysis run fails.

This change returns the Error stack which includes a helpful error
message and the backtrace for debugging purposes.