Adjusting the output table width to 80 chars
The default width for a standard terminal is 80 characters. cli-table2 colWidths only sets the width inside the cells without the boarders, and since there are two columns resulting in three boarders (left, middle, right), the sum for the width of the two cells need to add up to 77 for an 80 char wide table.